requirements.txt
defusedxml==0.5.0
Django==1.11
django-allauth==0.31.0
django-bootstrap-form==3.2.1
django-suit==0.2.25
facebook-sdk==2.0.0
facepy==1.0.8
oauthlib==2.0.2
PyJWT==1.4.2
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
Unidecode==0.4.20
